refactor: move chart permission checks into CoderService
